# Conflicts:
#	redrain-server/src/main/java/com/jredrain/job/RedRainTask.java
#	redrain-server/src/main/java/com/jredrain/service/ExecuteService.java

+++ redrain-server/src/main/java/com/jredrain/controller/JobController.java
@@ -159,7 +159,7 @@ public class JobController {
                 if (chind.getRedo() == 0) {
                     chind.setRunCount(null);
                 } else {
-                    chind.setRunCount(Long.parseLong((String) runCount[i]));
+                    chind.setRunCount(Integer.parseInt((String) runCount[i]));
                 }
                 chindren.add(chind);
             }
